fix(image-gen): extract parseSecondaryRuns to client-safe module (build fix)
CanvasContent.tsx (client component) imported parseSecondaryRuns from
layer-renderer.ts which has `import "server-only"`. This caused the
production build to fail with the documented server-only barrel leak pattern.

Fix: move parseSecondaryRuns + TextRun to lib/image/secondary-style-parser.ts
— no server-only, no I/O, pure string parsing. Both callers updated:
  - layer-renderer.ts: imports + re-exports from secondary-style-parser.ts
    (backward-compat for test imports)
  - CanvasContent.tsx: imports directly from secondary-style-parser.ts

Design spec §6.6/§1.7: "parser must be identical across all renderers."
The shared module ensures both the DOM renderer (client) and the sharp
renderer (server) use the exact same implementation.

Build: ✓ compiled. Tests: 273/273 pass.
